changed: reduce duplication in ASM 3D tests
use a common implementation where appropriate
