+++ CMakeLists.txt
@@ -620,6 +620,7 @@ if(TARGET IFEM-test)
         HEADERS
       FILES
         src/ASM/Test/ASM2DTests.h
+        src/ASM/Test/ASM3DTests.h
         src/ASM/Test/ASMCube.h
         src/ASM/Test/ASMSquare.h
         src/ASM/Test/ASMTrap.h
